requirementspip.txt
absl-py==2.1.0
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
archspec @ file:///croot/archspec_1709217642129/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
box2d-py==2.3.5
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
cloudpickle==3.1.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
conda @ file:///home/conda/feedstock_root/build_artifacts/conda_1729155160285/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///croot/conda-package-handling_1718138267740/work
conda_package_streaming @ file:///croot/conda-package-streaming_1718136078615/work
contourpy==1.3.0
cryptography @ file:///croot/cryptography_1724940562255/work
cycler==0.12.1
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1731044888992/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
distro @ file:///croot/distro_1714488253808/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
Farama-Notifications==0.0.4
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozendict @ file:///croot/frozendict_1713194832637/work
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
-e git+https://github.com/NotAnyMike/gym@794bcdf7f1e2a13c47451b6132b6e7ac626cf8d8#egg=gym
gym-notices==0.0.8
gymnasium==1.0.0
h5py==3.12.1
idna @ file:///croot/idna_1714398848350/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1729866374957/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1731317204262/work
Jinja2==3.1.4
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
libmambapy @ file:///croot/mamba-split_1725563367192/work/libmambapy
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1723567589013/work
ml-dtypes==0.4.1
mpmath==1.3.0
-e git+https://github.com/NotAnyMike/HRL@07446a734d3fbdef19da66d9c91dd80674caa9b3#egg=myproject
namex==0.0.8
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==11.0.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1727341649933/work
protobuf==5.28.3
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1729847057810/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pygame==2.6.1
pyglet==2.0.18
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.2.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz==2024.2
pyzmq @ file:///croot/pyzmq_1705605076900/work
requests @ file:///croot/requests_1721410876868/work
rich==13.9.4
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
scipy==1.14.1
setuptools==75.3.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stable_baselines3==2.4.0a11
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
swig==4.2.1.post0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-logger==0.1.0
tensorflow==2.18.0
termcolor==2.5.0
torch==2.5.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1724956131631/work
tqdm @ file:///croot/tqdm_1724853939799/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
triton==3.1.0
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.2
urllib3 @ file:///croot/urllib3_1727769808118/work
usd-core==24.8
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.16.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1731262100163/work
zstandard @ file:///croot/zstandard_1728569189425/work
